korzh_bot/requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
betterproto==1.2.5
chardet==3.0.4
grpclib==0.3.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pyasn1==0.4.8
pylint==2.5.3
python-dotenv==0.14.0
rsa==3.4.2
six==1.15.0
soupsieve==2.0.1
steamio==0.3.2
stringcase==1.2.0
toml==0.10.1
wrapt==1.12.1
yarl==1.4.2
